requirements.txt
aiogram==3.13.1
deepgram-sdk==3.7.7
anthropic==0.40.0
httpx==0.27.2
